Set thrown response headers in app layout boundary
